use shared truncateHashes
