+++ packages/auth/src/util/index.ts
@@ -8,5 +8,4 @@ export * from './composeTransforms.js'
 export * from './getScope.js'
 export * from './lockboxSummary.js'
 export * from './scopesMatch.js'
-export * from './truncateHashes.js'
 export * from './types.js'
