+++ packages/shared/src/index.ts
@@ -1,5 +1,6 @@
 export * from './assert.js'
 export * from './debug.js'
-export * from './pause.js'
 export * from './eventPromise.js'
 export * from './memoize.js'
+export * from './pause.js'
+export * from './truncateHashes.js'
